refactor(organization): reorganizes where components live, no logic changes
Moves lumberjack-specific components to the src/components folder, and general "this could be in any
project" components into pattern-library
