+++ client/src/pattern-library/components/TabbedContent/components/Tab/index.ts
