+++ client/src/pattern-library/components/TabbedContent/components/TabPanel/TabPanel.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { Box } from '../../../../pattern-library'
+import { Box } from '../../../../.'
 
 type PropsT = {
   children
