+++ client/src/pattern-library/components/TabbedContent/components/TabPanel/index.ts
