+++ client/src/pattern-library/components/TabbedContent/index.ts
